pyproject.toml
[project]
authors = [{ name = "Jakub Jezek", email = "jakubjezek001@gmail.com" }]
name = "jeza-ynput-dev-workspace"
version = "0.1.0"
description = "My personal workspace for YNPUT development."
readme = "README.md"
requires-python = ">=3.9.19"
dependencies = [
    "ayon-python-api>=1.0.11",
    "click>=8.1.7",
    "pyblish>=1.7.0",
    "python-dotenv>=1.0.1",
    "ruff>=0.8.3",
]

[project.scripts]
git-clone-all-repos = "jeza_ynput_dev_workspace:git_clone_all_repos"
upload-to-addon-folder = "jeza_ynput_dev_workspace:upload_to_addon_folder"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
venvPath = "."
venv = ".venv"

executionEnvironments = [
    { root = ".", extraPaths = [
        "ayon-backend",
        "ayon-applications/client",
        "ayon-blender/client",
        "ayon-core/client",
        "ayon-deadline/client",
        "ayon-houdini/client",
        "ayon-maya/client",
        "ayon-nuke/client",
        "ayon-ocio/client",
        "ayon-python-api",
        "ayon-third-party/client",
        "ayon-traypublisher/client",
        "example-studio-addon/client",
        "ayon-kitsu/client",
        "ayon-ftrack/client",
        "ayon-shotgrid/client",
        "ayon-circuit/client",
        "ayon-circuit/client/ayon_circuit/vendor/LabLib",
        "ayon-docker/backend",
        "ayon-launcher/.venv/lib64/python3.9/site-packages",
    ] },
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "W"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff]
line-length = 79
indent-width = 4

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
